Add multipart upload permissions to read-write policy
Add the required s3 permissions to the `read-write` policy to allow multi part file transfers. [1]

To allow the `write-only` policy to be compatible with multi part file transfers, these same actions/permissions are required. I did not add them in this commit because it's not completely clear if the `write-only` policy is compatible from a security perspective since the necessary permissions include `List*` actions.

1. https://docs.aws.amazon.com/AmazonS3/latest/userguide/mpuoverview.html